# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Allergy Scan API
  version: 1.0.0
  description: |
    アレルギースキャンアプリのバックエンド API。
    NestJS on AWS Lambda で実装。

    > このファイルは正典（source of truth）。API 変更時は Generator が spec-docs-syncer 経由で更新する。

servers:
  - url: https://api.example.com
    description: Production (AWS Lambda)
  - url: http://localhost:3001
    description: Local development

tags:
  - name: allergens
    description: アレルゲンマスター
  - name: scan
    description: スキャン（バーコード・OCR）
  - name: history
    description: スキャン履歴
  - name: users
    description: ユーザー設定

paths:
  /users/init:
    post:
      tags: [users]
      summary: 初回 Cookie 発行・ユーザー作成
      description: |
        Cookie が未設定の場合は UUID を生成して users テーブルに INSERT し Cookie を発行する。
        既存 Cookie がある場合はそのまま onboarding_done を返す。
      responses:
        "200":
          description: 既存ユーザー（Cookie あり）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitResponse"
              example:
                created: false
                onboarding_done: true
        "201":
          description: 新規ユーザー作成・Cookie 発行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitResponse"
              example:
                created: true
                onboarding_done: false

  /users/backup-code:
    post:
      tags: [users]
      summary: バックアップコード発行
      description: |
        デバイス引き継ぎ用コードを発行する。Cookie 認証必須。
        既存の未使用コードは失効させてから新コードを発行する。
        有効期限: 7日。フォーマット: ALRG-XXXX-XXXX（文字セットから O/0/I/1 を除外）。
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupCodeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /users/restore:
    post:
      tags: [users]
      summary: バックアップコードでデバイス引き継ぎ
      description: |
        バックアップコードを照合して scan_histories・allergies を新デバイスに移行する。
        レートリミット: 60秒5回（ブルートフォース防止）。
        処理はトランザクション内で実行。
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "200":
          description: 引き継ぎ成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Too Many Requests（レートリミット）

  /products/others:
    get:
      tags: [scan]
      summary: みんなのスキャン一覧
      description: |
        リクエストユーザーがスキャンしていない商品一覧をカーソルページネーションで返す。
        Cookie 認証必須。ユーザーのアレルゲン設定に基づいた judgment を付与する。
      security:
        - cookieAuth: []
      parameters:
        - name: cursor
          in: query
          description: 前ページの最後の updated_at（ISO 8601）
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OthersProductListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /allergens:
    get:
      tags: [allergens]
      summary: アレルゲンマスター取得
      description: 設定画面用。category・display_order 順で返す。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllergenCategory"
              example:
                - category: mandatory
                  label: "特定原材料（9品目・表示義務あり）"
                  items:
                    - name: えび
                      emoji: 🦐
                      display_order: 1
                    - name: 卵
                      emoji: 🥚
                      display_order: 7
                - category: recommended
                  label: "準ずるもの（20品目・表示推奨）"
                  items:
                    - name: アーモンド
                      emoji: 🌰
                      display_order: 10

  /scan/presigned-url:
    get:
      tags: [scan]
      summary: S3 Presigned URL 発行
      description: |
        クライアントが直接 S3 に画像をアップロードするための URL を発行する。
        Lambda の 6MB ペイロード制限を回避するため、画像は Lambda 経由で受け取らない。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUrlResponse"

  /scan/barcode:
    post:
      tags: [scan]
      summary: バーコード（JAN コード）照合
      description: |
        照合フロー:
        1. NestJS メモリキャッシュ確認（TTL: 60秒）
        2. DB の expires_at 確認（有効期限内なら即返却）
        3. Open Food Facts API 照合
        4. ミスなら `found: false` を返却（クライアントが OCR に自動切り替え）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarcodeScanRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarcodeScanResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /scan/ocr:
    post:
      tags: [scan]
      summary: OCR + アレルゲン判定
      description: |
        フロー:
        1. S3 から画像取得
        2. ユーザーの有効アレルゲンを取得
        3. allergen_components から成分リストを動的生成
        4. Gemini Flash API に送信
        5. incomplete: true → 400 返却
        6. confidence: low → 422 返却（再スキャン誘導）
        7. products テーブルに UPSERT（scan_count +1、expires_at 再計算）
        8. scan_histories に記録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrScanRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrScanResponse"
        "400":
          description: incomplete（ラベル全体が映っていない）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: ラベル全体が映るように離してください
                code: INCOMPLETE_IMAGE
        "422":
          description: confidence が low（近づけて再スキャン）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: もう少し近づけて再スキャンしてください
                code: LOW_CONFIDENCE

  /history:
    get:
      tags: [history]
      summary: スキャン履歴一覧
      description: カーソルベースページネーション。20件ずつ返す。
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, ng, partial, ok]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: cursor
          in: query
          description: 前ページの最後の scanned_at（ISO 8601）
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"

    post:
      tags: [history]
      summary: スキャン履歴保存
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistorySaveRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanHistory"
        "400":
          $ref: "#/components/responses/BadRequest"

  /history/{id}:
    patch:
      tags: [history]
      summary: スキャン履歴の location 更新
      description: |
        店舗候補が 2 件以上の場合にユーザーが選択した店舗を後から設定するエンドポイント。
        Cookie の userId と history.user_id が一致しない場合は 403 を返す。
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchHistoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden（他ユーザーの履歴への操作）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      tags: [users]
      summary: ユーザー設定取得
      description: TTL 5分でキャッシュ。スキャン開始時に呼び出し Gemini プロンプト動的生成に使用。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"

    put:
      tags: [users]
      summary: アレルギー設定更新
      description: |
        enabled を true にした瞬間、partialAlert も自動で true になる。
        enabled を false にした瞬間、partialAlert も自動で false になる。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllergySettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          $ref: "#/components/responses/BadRequest"

    delete:
      tags: [users]
      summary: ユーザーデータ削除
      description: |
        要配慮個人情報（アレルギー設定・スキャン履歴）を全削除。
        個人情報保護法上、ユーザーからの削除要求に対応するため必須。
      responses:
        "204":
          description: No Content

components:
  schemas:
    AllergenCategory:
      type: object
      required: [category, label, items]
      properties:
        category:
          type: string
          enum: [mandatory, recommended]
        label:
          type: string
          example: "特定原材料（9品目・表示義務あり）"
        items:
          type: array
          items:
            $ref: "#/components/schemas/AllergenItem"

    AllergenItem:
      type: object
      required: [name, emoji, display_order]
      properties:
        name:
          type: string
          example: 乳
        emoji:
          type: string
          example: 🥛
        display_order:
          type: integer
          example: 8

    AllergenSetting:
      type: object
      required: [enabled, partialAlert]
      properties:
        enabled:
          type: boolean
        partialAlert:
          type: boolean

    PresignedUrlResponse:
      type: object
      required: [url, s3_key]
      properties:
        url:
          type: string
          format: uri
          description: S3 PUT 用 Presigned URL
        s3_key:
          type: string
          description: OCR リクエスト時に渡す S3 キー

    BarcodeScanRequest:
      type: object
      required: [jan_code]
      properties:
        jan_code:
          type: string
          pattern: '^\d{8,13}$'
          example: "4901234567890"

    BarcodeScanResponse:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        product_name:
          type: string
          nullable: true
        allergens:
          $ref: "#/components/schemas/ProductAllergens"
          nullable: true
        judgment:
          $ref: "#/components/schemas/Judgment"
          nullable: true
        detected:
          type: array
          items:
            type: string
          nullable: true
        from_cache:
          type: boolean

    ProductAllergens:
      type: object
      properties:
        contains:
          type: array
          items:
            type: string
          description: 含むアレルゲン
        partial:
          type: array
          items:
            type: string
          description: 一部含むアレルゲン
        components:
          type: array
          items:
            type: string
          description: 検出された派生成分

    OcrScanRequest:
      type: object
      required: [s3_key]
      properties:
        s3_key:
          type: string
          description: /scan/presigned-url で取得した S3 キー

    OcrScanResponse:
      type: object
      required: [raw_text, confidence, results, highlights, incomplete]
      properties:
        raw_text:
          type: string
          description: OCR で読み取った原材料テキスト全文
        confidence:
          $ref: "#/components/schemas/Confidence"
        results:
          type: array
          items:
            $ref: "#/components/schemas/AllergenResult"
          description: アレルゲンごとの判定結果（複数）
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/HighlightItem"
          description: UI ハイライト用テキスト一覧（raw_text 内の検出箇所）
        incomplete:
          type: boolean
          description: ラベル全体が映っていない場合 true
        price:
          type: number
          nullable: true
          description: 税抜価格（取得できた場合のみ）
        price_with_tax:
          type: number
          nullable: true
          description: 税込価格
        price_confidence:
          type: string
          enum: [high, low]
          nullable: true
          description: 価格の読み取り信頼度（high のみ表示）

    AllergenResult:
      type: object
      required: [allergen, judgment, detection_type, detected, risk_level, reason]
      properties:
        allergen:
          type: string
          description: アレルゲン名（allergens.name と一致）
        judgment:
          $ref: "#/components/schemas/Judgment"
        detection_type:
          type: string
          enum: [contains, partial, may_contain]
          description: |
            contains: 原材料として直接含む（🔴 NG）
            partial: 一部に含む・一括表示パターン（🟡 注意）
            may_contain: 製造ラインのコンタミ注意喚起（🟠 注意喚起）
        detected:
          type: array
          items:
            type: string
          description: 検出された成分名リスト
        risk_level:
          type: string
          enum: [high, medium, low, ignore]
          description: アナフィラキシーリスクレベル
        reason:
          type: string
          description: 判定理由

    HighlightItem:
      type: object
      required: [text, judgment]
      properties:
        text:
          type: string
          description: raw_text 内のハイライト対象テキスト
        judgment:
          type: string
          enum: [ng, partial, may_contain]
          description: "ng: 🔴 / partial: 🟡 / may_contain: 🟠"

    HistoryListResponse:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanHistory"
        nextCursor:
          type: string
          format: date-time
          nullable: true

    ScanHistory:
      type: object
      required: [id, product_name, judgment, scanned_at]
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
          nullable: true
        judgment:
          $ref: "#/components/schemas/JudgmentShort"
        detected:
          type: array
          items:
            type: string
        location:
          $ref: "#/components/schemas/Location"
          nullable: true
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          description: 惣菜のみ（S3 URL）
        price:
          type: number
          nullable: true
        scanned_at:
          type: string
          format: date-time

    HistorySaveRequest:
      type: object
      required: [product_id, judgment]
      properties:
        product_id:
          type: string
          format: uuid
          nullable: true
        product_name:
          type: string
          nullable: true
        judgment:
          $ref: "#/components/schemas/JudgmentShort"
        detected:
          type: array
          items:
            type: string
        location:
          $ref: "#/components/schemas/Location"
          nullable: true
        thumbnail_url:
          type: string
          format: uri
          nullable: true

    PatchHistoryRequest:
      type: object
      required: [location]
      properties:
        location:
          $ref: "#/components/schemas/Location"

    UserSettings:
      type: object
      required: [id, allergies, locale, onboarding_done]
      properties:
        id:
          type: string
          description: デバイス ID
        allergies:
          type: object
          description: キーは allergens.name と完全一致
          additionalProperties:
            $ref: "#/components/schemas/AllergenSetting"
          example:
            乳:
              enabled: true
              partialAlert: true
            卵:
              enabled: false
              partialAlert: false
        locale:
          type: string
          enum: [ja, en]
          description: ユーザーの言語設定
        onboarding_done:
          type: boolean
          description: オンボーディング完了フラグ

    AllergySettingsUpdate:
      type: object
      properties:
        allergies:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AllergenSetting"
        locale:
          type: string
          enum: [ja, en]
        onboarding_done:
          type: boolean
          description: true のみ受け付ける（false への上書きはサーバー側で無視される）

    Location:
      type: object
      properties:
        store_name:
          type: string
          nullable: true
          example: セブンイレブン渋谷店
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true

    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string

    InitResponse:
      type: object
      required: [created, onboarding_done]
      properties:
        created:
          type: boolean
          description: true = 新規ユーザー作成、false = 既存ユーザー
        onboarding_done:
          type: boolean

    BackupCodeResponse:
      type: object
      required: [code, expires_at]
      properties:
        code:
          type: string
          pattern: '^ALRG-[A-HJ-NP-Z2-9]{4}-[A-HJ-NP-Z2-9]{4}$'
          example: ALRG-4829-KXMZ
        expires_at:
          type: string
          format: date-time
          description: 有効期限（発行から7日後）

    RestoreRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: '^ALRG-[A-HJ-NP-Z2-9]{4}-[A-HJ-NP-Z2-9]{4}$'

    RestoreResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true

    OthersProductListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OthersProductItem"
        next_cursor:
          type: string
          format: date-time
          nullable: true

    OthersProductItem:
      type: object
      required: [id, allergens, judgment, detected, updated_at, is_expired]
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
          nullable: true
        allergens:
          $ref: "#/components/schemas/ProductAllergens"
        judgment:
          $ref: "#/components/schemas/JudgmentShort"
          description: ユーザーのアレルゲン設定に基づいた総合判定
        detected:
          type: array
          items:
            type: string
          description: ng/partial 時に検出されたアレルゲン名
        updated_at:
          type: string
          format: date-time
        is_expired:
          type: boolean
          description: expires_at < NOW() の場合 true

    Judgment:
      type: string
      enum: [含む, 一部含む, なし, 判定不能]

    JudgmentShort:
      type: string
      enum: [ng, partial, ok]
      description: ng=含む / partial=一部含む / ok=なし

    Confidence:
      type: string
      enum: [high, medium, low]

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: user_id

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized（Cookie なし）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"